# Conflicts:
#	src/main/kotlin/org/example/coha/domain/post/controller/PostController.kt
#	src/main/kotlin/org/example/coha/domain/post/dto/UpdatePostRequest.kt
#	src/main/kotlin/org/example/coha/domain/post/model/Post.kt
#	src/main/kotlin/org/example/coha/domain/post/service/PostService.kt
#	src/main/kotlin/org/example/coha/domain/post/service/PostServiceImpl.kt
#	src/main/kotlin/org/example/coha/domain/reply/controller/ReplyController.kt
#	src/main/kotlin/org/example/coha/domain/reply/service/ReplyService.kt
#	src/main/kotlin/org/example/coha/domain/reply/service/ReplyServiceImpl.kt

+++ src/main/kotlin/org/example/coha/domain/post/service/PostServiceImpl.kt
@@ -1,12 +1,14 @@
 package org.example.coha.domain.post.service
 
 import org.example.coha.domain.exception.ModelNotFoundException
+import org.example.coha.domain.exception.UnauthorizedAccess
 import org.example.coha.domain.post.dto.CreatePostRequest
 import org.example.coha.domain.post.dto.PostResponse
 import org.example.coha.domain.post.dto.PostWithReplyResponse
 import org.example.coha.domain.post.dto.UpdatePostRequest
 import org.example.coha.domain.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -16,8 +18,11 @@ class PostServiceImpl(
     private val postRepository: PostRepository
 ): PostService {
 
+    @Transactional
     override fun createPost(request: CreatePostRequest): PostResponse {
-        val post = postRepository.save(request.toPost())
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+
+        val post = postRepository.save(request.toPost(currentUser))
         return PostResponse.toPostResponse(post)
 
 
@@ -38,15 +43,22 @@ class PostServiceImpl(
     }
 
     @Transactional
-    override fun updatePost(postId: Long): PostResponse {
-        val savedpost = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
-        savedpost.content
-        return PostResponse.toPostResponse(savedpost)
+
+    override fun updatePost(postId: Long, request: UpdatePostRequest): PostResponse {
+        val savedPost = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+        if(savedPost.author != currentUser) throw UnauthorizedAccess()
+        savedPost.updatePost(request)
+        return PostResponse.toPostResponse(savedPost)
+
     }
 
 
     @Transactional
     override fun deletePost(postId: Long) {
+        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+        if(post.author != currentUser) throw UnauthorizedAccess()
         postRepository.deleteById(postId)
     }
     @Transactional
@@ -54,4 +66,10 @@ class PostServiceImpl(
         postRepository.updateViews(postId)
     }
 
+
+
+
 }
+
+
+
